+++ changelog.md
@@ -1,3 +1,12 @@
+0.2.0 (2018-09-24)
+==================
+
+* Expire sessions.
+* Add garbage collector to MemoryStore.
+* BC break: Now uses `ctx.state.session` and `ctx.state.sessionId` instead of
+  `ctx.state.session.data` and `ctx.state.session.id`.
+
+
 0.1.1 (2018-09-06)
 ==================